fix:修改套餐部分接口报错

+++ mazongjian-module-member/mazongjian-module-member-biz/src/main/java/com/yanzu/module/member/controller/app/pkg/vo/AppAdminPkgSaveReqVO.java
@@ -38,7 +38,7 @@ public class AppAdminPkgSaveReqVO {
     private String ksId;
 
     @Schema(description = "按包厢设置")
-    private List<Integer> enableRoom;
+    private List<Long> enableRoom;
 
     @Schema(description = "可用时间 0-23数字，空表示不限")
     private List<Integer> enableTime;
